pyproject.toml
[tool.poetry]
name = "ShigureBot"
version = "0.1.0"
description = "Kawaii no Shigure Bot~"
authors = ["student_2333 <lgc2333@126.com>"]
license = "GPL-3.0 License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<4.0.0"
nonebot2 = "^2.0.0b2"
aiohttp = "^3.8.1"
aiofiles = "^0.8.0"
mcstatus = "^9.0.3"
pydantic = "^1.9.0"
nonebot-plugin-htmlrender = "^0.0.4.5"
nonebot-plugin-apscheduler = "^0.1.2"
playwright = "^1.20.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.6"

[nonebot.plugins]
plugins = []
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"